// haredube/src/modules/websocket/websocket.enum.ts
export enum SOCKET_SCREEN {
  MEETING = 'MEETING',
}

export enum SOCKET_SCREEN_ACTION {
  JOIN = 'JOIN',
  LEAVE = 'LEAVE',
  CREATE_METTING = 'CREATE_METTING',
  USER_JOINED = 'USER_JOINED',
  USER_LEFT = 'USER_LEFT',
  SIGNAL = 'SIGNAL',
  DISCONNECT = 'DISCONNECT',
  MESSAGE = 'MESSAGE',
  JOIN_NOTIFY = 'JOIN_NOTIFY',
  SEND_NOTIFY = 'SEND_NOTIFY',
  GET_NOTIFY = 'GET_NOTIFY',
  USER_JOIN_NOTIFY = 'USER_JOIN_NOTIFY',
}